Updated 'phantomjs' path to absolute path
Struggled for a while trying to get this to schedule properly.  I could run the script fine manually through terminal but when crontab scheduling, it struggled to find phantomjs.  Changing the launcher to use the full path '/usr/local/bin/phantomjs' instead of just 'phantomjs' fixed my issue.  If you feel this was just my own local mac issue and don't think this commit is necessary, it might be worth at least adding some help documentation on trying to use 'usr/local/bin/phantomjs' instead of 'phantomjs' due to the amount of people attempting to use this that aren't tech-savvy.