My custom tweaks
Thanks so much for this project, just got four appointments for my entire family with it <3

I tweaked a few things so I figured I would send a PR to give some ideas. I don't expect you to take any of it as is but it may be inspiration.

Instead of opening up an OSX notification it opens a new tab in chrome with the pretty printed day. I've also added a blacklist with days so that it doesn't repeatedly open the tab as I'm doing it on a while loop.